requirements.txt
greenlet==3.0.3
numpy==2.1.1
pandas==2.2.2
pyodbc==5.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.33
typing_extensions==4.12.2
tzdata==2024.1
